Updated README.md
As I was adding the class repo as an upstream repo, I noticed that the 'master' branch was actually titled 'main'. Apparently, GitHub recently decided to dump the term 'master' for a more neutral term. The changes in this README replace the instances of anything having to do with an upstream 'master' branch with the correct 'main' branch title. Also, I quickly want to piggy-back off of what Teyler said in his pull request; you are an awesome professor. I am thoroughly looking forward to taking this class. Thank you!